PRSD-422: Moves enums package into constants package
