+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/constants/enums/JourneyType.kt
@@ -1,0 +1,5 @@
+package uk.gov.communities.prsdb.webapp.constants.enums
+
+enum class JourneyType {
+    LANDLORD_REGISTRATION,
+}
