--- src/main/kotlin/uk/gov/communities/prsdb/webapp/enums/JourneyType.kt
@@ -1,5 +1,0 @@
-package uk.gov.communities.prsdb.webapp.enums
-
-enum class JourneyType {
-    LANDLORD_REGISTRATION,
-}
